move test utilities into root
