+++ packages/browser-destinations/src/destinations/logrocket/__tests__/index.test.ts
@@ -1,7 +1,6 @@
 import { Analytics, Context } from '@segment/analytics-next'
 import plugins, { destination } from '../index'
-import { subscriptions } from './subscriptions'
-import { mockWorkerAndXMLHttpRequest } from './utilities'
+import { mockWorkerAndXMLHttpRequest, subscriptions } from '../test_utilities'
 
 describe('Logrocket', () => {
   beforeAll(mockWorkerAndXMLHttpRequest)
